Update processing of resource files

+++ src/tools/File.h
@@ -46,9 +46,11 @@ public:
     bool IsOpen() const { return m_stream.is_open(); }
 
     /// @brief Getting the current position
+    /// @return start position for processing
     std::streampos Position() { return m_stream.tellg(); }
 
     /// @brief Getting flag about end of file
+    /// @return true if file has data for reading
     bool IsEof() const { return m_stream.eof(); }
 
     /// @brief Setting the absolute position
@@ -60,22 +62,26 @@ public:
     void Seek(std::streamoff offset)  { m_stream.seekg(offset, std::ios_base::cur); }
 
     /// @brief Getting a byte from file
-    uint8_t  Get() { return m_stream.get(); }
+    /// @return reading byte
+    uint8_t  Get() { return static_cast<uint8_t>(m_stream.get()); }
 
     /// @brief Getting two byte from file
+    /// @return reading word
     uint16_t Get16LE() { uint16_t value; m_stream.read(reinterpret_cast<char *>(&value), 2); return value; }
 
     /// @brief Getting three byte from file
+    /// @return reading double bytes, which has value in three low bytes
     uint32_t Get24LE() { uint32_t value; m_stream.read(reinterpret_cast<char *>(&value), 3); return value; }
 
     /// @brief Getting four byte from file
+    /// @return reading double word
     uint32_t Get32LE() { uint32_t value; m_stream.read(reinterpret_cast<char *>(&value), 4); return value; }
 
     /// @brief Read data from file
     /// @param buffer   - container for data
     /// @param size     - size of container
     /// @param err      - standard stream for output message about error
-    /// @return true if reading was success
+    /// @return true if reading is successed
     bool Read(Data &buffer, Data::size_type size, std::ostream &err) noexcept;
 
 /*
